Move queryTile method to backend and proxy on vector source.
