+++ index.js
@@ -286,31 +286,7 @@ Vector.prototype.getInfo = function(callback) {
 // letting the tilelive-vector backend do the hard work of finding
 // the right tile to use.
 Vector.prototype.queryTile = function(z, lon, lat, options, callback) {
-    var xyz = sm.xyz([lon, lat, lon, lat], z);
-    this._backend.getTile(z, xyz.minX, xyz.minY, function(err, vtile, head) {
-        if (err) return callback(err);
-        try {
-            var features = vtile.query(lon, lat, options);
-        } catch(err) {
-            return callback(err);
-        }
-        var results = [];
-        for (var i = 0; i < features.length; i++) {
-            results.push({
-                id: features[i].id(),
-                distance: features[i].distance,
-                layer: features[i].layer,
-                attributes: features[i].attributes()
-            });
-        }
-        var headers = {};
-        headers['Content-Type'] = 'application/json';
-        headers['ETag'] = JSON.stringify(crypto.createHash('md5')
-            .update(head && head['ETag'] || (z+','+lon+','+lat))
-            .digest('hex'));
-        headers['Last-Modified'] = new Date(head && head['Last-Modified'] || 0).toUTCString();
-        return callback(null, results, headers);
-    });
+    this._backend.queryTile(z, lon, lat, options, callback);
 };
 
 Vector.prototype.profile = function(callback) {
